Completed the refactor
