+++ resolvers/SingleItemLookup/PersonResolver.js
@@ -1,14 +1,14 @@
 const axios = require("axios");
 const { find } = require("lodash");
+
 const {
   generateSearchEndpoint,
-  generateSingleItemLookupEndpoint,
-  generateImageURL
-} = require("../../config");
+  generateSingleItemLookupEndpoint
+} = require("../../utils/generateEndpoints");
 
 const SearchForAPersonResolver = async (parent, args, context, info) => {
   try {
-    // 1. Make a request to the search API using a search term provided in the query
+    // Make a request to the search API using a search term provided in the query
     const response = await axios.get(
       generateSearchEndpoint(args.search, "person")
     );
@@ -16,11 +16,11 @@ const SearchForAPersonResolver = async (parent, args, context, info) => {
     const { data } = response;
     const { results } = data;
 
-    // 2. Find a person from the search results
-    const SinglePerson = await find(results, person => person.id === args.id);
+    // Find a person from the search results
+    const SinglePerson = find(results, person => person.id === args.id);
 
-    // 3. Perform a single person lookup using the person found in search results array
     try {
+      // Perform a single person lookup using the person found in search results array
       const response = await axios.get(
         generateSingleItemLookupEndpoint(SinglePerson.id, "person")
       );
@@ -29,13 +29,11 @@ const SearchForAPersonResolver = async (parent, args, context, info) => {
       return data;
     } catch (err) {
       console.log(`The /Person endpoint failed`);
-      console.log(err);
-      return err;
+      return err.data;
     }
   } catch (err) {
     console.log("The /Search endpoint failed");
-    console.log(err);
-    return err;
+    return err.data;
   }
 };
