+++ resolvers/Social/ShowResolver.js
@@ -1,8 +1,12 @@
 const axios = require("axios");
-const { generateSocialLinksEndpoint } = require("../../config");
+
+const {
+  generateSocialLinksEndpoint
+} = require("../../utils/generateEndpoints");
 
 const TVSocialResolver = async (parent, args, info, context) => {
   try {
+    // Make a request to the social (external_links) endpoint using the parent id
     const response = await axios.get(
       generateSocialLinksEndpoint(parent.id, "tv")
     );
@@ -12,8 +16,7 @@ const TVSocialResolver = async (parent, args, info, context) => {
     return data;
   } catch (err) {
     console.log("The TV /external_ids (social) endpoint failed");
-    console.log(err);
-    return err;
+    return err.data;
   }
 };
