+++ resolvers/index.js
@@ -62,7 +62,7 @@ const { MovieVideoResolver, ShowVideoResolver } = require("./Videos");
 const { MovieCreditsResolver, ShowCreditsResolver } = require("./Credits");
 
 const resolvers = {
-  // Additional show data lookup (Show refers to the model, the single person lookup id is used to retriveve the additional data)
+  // Additional data for the single movie object
   Movie: {
     reviews: MovieReviewResolver,
     cast: MovieCastResolver,
@@ -73,7 +73,7 @@ const resolvers = {
     videos: MovieVideoResolver
   },
 
-  // Additional show data lookup (Show refers to the model, the single show lookup id is used to retriveve the additional data)
+  // Additional data for the single show object
   Show: {
     reviews: ShowReviewResolver,
     cast: ShowCastResolver,
@@ -84,11 +84,13 @@ const resolvers = {
     videos: ShowVideoResolver
   },
 
+  // Additional data for the single person object
   Person: {
     tv_credits: ShowCreditsResolver,
     movie_credits: MovieCreditsResolver
   },
 
+  // Root query
   Query: {
     // SingleItemLookup
     SearchForAMovie: SearchForAMovieResolver,
